IMPORTANT: remove Grid interp, k-less parts of code
This removes the `interpolatev_G` functions from Interpolations.jl, as
well as the function taking `CoordInterpGrid` as inputs from Bellman.jl
and Optimizations.jl.
Further, the functions not using `k` (the Kalman gain) are removed from
Optimizations.jl.
The `solveWorkingLife` method not using `k` is removed, and the methods
for constant and time-varying belief grids are merged into one.